feature(typescript): extracted all types from `index.d.ts` into individual `.d.ts` files
