+++ src/components/code-pane.d.ts
@@ -1,0 +1,10 @@
+import * as React from 'react';
+
+export const CodePane: React.FC<{
+  children: React.ReactNode;
+  language: string;
+  theme?: Record<string, unknown>;
+  stepIndex?: number;
+  highlightRanges?: number | (number | number[])[];
+  showLineNumbers?: boolean;
+}>;
