+++ src/components/fullscreen.d.ts
@@ -1,0 +1,6 @@
+import * as React from 'react';
+
+export const FullScreen: React.FC<{
+  color: string;
+  size: number;
+}>;
