+++ src/components/logo.d.ts
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+export const SpectacleLogo: React.FC<{
+  size: number;
+}>;
