+++ src/components/notes.d.ts
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+export const Notes: React.FC<{
+  children: React.ReactNode;
+}>;
