+++ src/components/progress.d.ts
@@ -1,0 +1,6 @@
+import * as React from 'react';
+
+export const Progress: React.FC<{
+  color: string;
+  size: number;
+}>;
