+++ src/components/transitions/index.d.ts
@@ -1,0 +1,9 @@
+export type SlideTransition = {
+  from: Record<string, string | number>;
+  leave: Record<string, string | number>;
+  enter: Record<string, string | number>;
+};
+
+export const fadeTransition: SlideTransition;
+export const slideTransition: SlideTransition;
+export const defaultTransition: SlideTransition;
