+++ src/utils/indent-normalizer.test.ts
@@ -2,6 +2,7 @@ import indentNormalizer from './indent-normalizer';
 
 describe('indentNormalizer', () => {
   it('handles empty cases', () => {
+    // @ts-ignore
     expect(indentNormalizer()).toEqual('');
     expect(indentNormalizer(null)).toEqual('');
     expect(indentNormalizer('')).toEqual('');
